ItemVIew

+++ hello-shop/hello-shop/src/main/java/shop/helloshop/domain/entity/items/Item.java
@@ -30,7 +30,7 @@ public abstract class Item {
 
     private LocalDateTime localDateTime;
 
-    private Long salesQuantity;
+    private int salesQuantity;
 
     @OneToMany(mappedBy = "item",cascade = CascadeType.ALL )
     private List<UploadFile> uploadFiles = new ArrayList<>();
